+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/preparator/extension/EncryptThenMacExtensionPreparatorTest.java
@@ -37,6 +37,6 @@ public class EncryptThenMacExtensionPreparatorTest {
         preparator.prepare();
 
         assertArrayEquals(extensionType.getValue(), message.getExtensionType().getValue());
-        assertEquals(extensionLength, (int) message.getExtensionLength().getValue());
+        assertEquals(extensionLength, (long) message.getExtensionLength().getValue());
     }
 }
